Contact Form: Updates to data exporter/eraser (#9638)
* Contact Form: Use associative keys for data exporters/erasers

In [#43931-core](https://core.trac.wordpress.org/ticket/43931) the decision was made to allow personal data exporters and erasers to be registered using associative key strings. This makes it easier to change what is in those lists of active callbacks because you don't have to iterate over the entire array to find what you're looking for.

This commit adds `jetpack-contact-form` as the key string for both the exporter and the eraser.

* Contact Form: Return booleans in eraser callback

In `personal_data_eraser`, the returned array includes items for
`items_removed` and `items_retained`. In Core, these were originally
intended to contain an integer count of the number of items for each
outcome. However,
in [r43061](https://core.trac.wordpress.org/changeset/43061), this
was changed so they just contained booleans instead.

[#43602-core](https://core.trac.wordpress.org/ticket/43602) has some
conversation that kind of explains why this change was made, but it's
a bit hard to follow the thread.

This commit brings Grunion's `personal_data_eraser` method in line
with this change, by switching from integers to booleans.

* Contact Form: Add filter to allow erasure prevention

This commit adds the `grunion_contact_form_delete_feedback_post` filter
which allows individual items to be skipped during erasure, similar to
how Core's comments eraser works. The default value of `true` can be
changed to a message string that explains why an item was not erased.

* Contact Form: Change exporter/eraser ID strings

Per discussion with @jaswrks, change `contact-form` to `feedback` for the exporter and eraser array keys. This more closely relates those callbacks to the type of data they are working with.

* Contact Form: Finish incomplete filter doc block

Adds a best-guess version number for the `@since` property and gives a variable name to the first `@param`.

+++ modules/contact-form/grunion-contact-form.php
@@ -749,7 +749,7 @@ class Grunion_Contact_Form_Plugin {
 	 * @return array $exporters An array of personal data exporters.
 	 */
 	public function register_personal_data_exporter( $exporters ) {
-		$exporters[] = array(
+		$exporters['jetpack-feedback'] = array(
 			'exporter_friendly_name' => __( 'Feedback', 'jetpack' ),
 			'callback'               => array( $this, 'personal_data_exporter' ),
 		);
@@ -767,7 +767,7 @@ class Grunion_Contact_Form_Plugin {
 	 * @return array $erasers An array of personal data erasers.
 	 */
 	public function register_personal_data_eraser( $erasers ) {
-		$erasers[] = array(
+		$erasers['jetpack-feedback'] = array(
 			'eraser_friendly_name' => __( 'Feedback', 'jetpack' ),
 			'callback'             => array( $this, 'personal_data_eraser' ),
 		);
@@ -844,16 +844,43 @@ class Grunion_Contact_Form_Plugin {
 	 */
 	public function personal_data_eraser( $email, $page = 1 ) {
 		$per_page = 250;
-		$removed  = 0;
-		$retained = 0;
+		$removed  = false;
+		$retained = false;
 		$messages = array();
 		$post_ids = $this->personal_data_post_ids_by_email( $email, $per_page, $page );
 
 		foreach ( $post_ids as $post_id ) {
+			/**
+			 * Filters whether to erase a particular Feedback post.
+			 *
+			 * @since 6.3.0
+			 *
+			 * @param bool|string $prevention_message Whether to apply erase the Feedback post (bool).
+			 *                                        Custom prevention message (string). Default true.
+			 * @param int         $post_id            Feedback post ID.
+			 */
+			$prevention_message = apply_filters( 'grunion_contact_form_delete_feedback_post', true, $post_id );
+
+			if ( true !== $prevention_message ) {
+				if ( $prevention_message && is_string( $prevention_message ) ) {
+					$messages[] = esc_html( $prevention_message );
+				} else {
+					$messages[] = sprintf(
+						// translators: %d: Post ID.
+						__( 'Feedback ID %d could not be removed at this time.', 'jetpack' ),
+						$post_id
+					);
+				}
+
+				$retained = true;
+
+				continue;
+			}
+
 			if ( wp_delete_post( $post_id, true ) ) {
-				$removed++;
+				$removed = true;
 			} else {
-				$retained++;
+				$retained = true;
 				$messages[] = sprintf(
 					// translators: %d: Post ID.
 					__( 'Feedback ID %d could not be removed at this time.', 'jetpack' ),
